tracing_agents: atrace: Stop trace before async dump
For long traces, async_dump can take a long time. During this time the tracer
is still running. Stop the tracer before collecting the dump so that the traces
are properly generated.

Signed-off-by: Joel Fernandes <joelaf@google.com>

devil/adb: Increase timeout of commands to 5 minutes

Some commands can take very long time to run, such as when collecting GB of
traces. Increase the timeout period.

Signed-off-by: Joel Fernandes <joelaf@google.com>
Bug: catapult:#
Change-Id: Ic0cec60ebf94c87b2a0a2b9d53030a9a6b70a039