Dashboard - Add pinpoint-perf-job-dialog for perf try jobs.
Dashboard initiates perf try jobs when you hit "request debug trace" via the chart tooltip. This adds a Pinpoint version of the trace dialog that routes the request to Pinpoint. Basically cloned the trace-form dialog into pinpoint-perf-job-dialog, and changed the internals to communicate with the /pinpoint endpoint. Should look and act exactly the same, with the exception that it will say "Send to Pinpoint".

Bug: chromium:768530
Change-Id: I65d91824a1f73436460ee39e68ce54946d2b279b