Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I don't think we want to edit these defaults, we should set our own timeout.
> 
> Can you edit 'devil_dynamic_config' in run_systrace.py to configure these? Or just pass the timeout to the individual places needed?
> 
> +jbudorick

There isn't a way to pass the timeout AFAIK. about devil_dynamic_config is that something that will be set for all users? If not I am not happy with keeping the tool broken for trace features that need a timeout. The experience for the user IMO should be automatic, not one where they have to edit a config to fix an error.

Patch-set: 1
Reviewer: Gerrit User 1245231 <1245231@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>