Sort overview chart in histogram-set-table-name-cell.
Currently, the overview line chart in the Name column cells is generated from a Map
from column title to Histogram. However, Maps are not guaranteed to be in any
particular order. This CL adds a generator method to HistogramSetTableRow that
yields the column titles and Histograms in the same order as the baseTable's
columns.

Before:
http://i.imgur.com/yHwbhNI.png

After:
http://i.imgur.com/BFYjKWg.png

Change-Id: I6dc294d3339949a45fe2283375f478da6c402ed3