Show %Δavg instead of Δavg.
In results.html, when the user selects a reference column, it automatically
switches to displaying delta statistics in the other columns.
However, interpreting Δavg frequently requires significant familiarity with the
metric. The percent delta version of the statistic (e.g. %Δavg) would probably
be easier to interpret for more users.

This CL just adds the '%' character.

http://www/~benjhayden/691275.html

https://github.com/catapult-project/catapult/issues/3911

Change-Id: Ia1ae9f1eeb6bec2c0e3fd6547b1b2f874b7c6d5f