Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Agree dynamic doesn't make sense.
> 
> Would still appreciate a datapoint or two for why 300 was chosen in a comment/the commit message.

Sure, it took around 180 seconds to grab the trace, I chose 300 just to be safe incase in the future our traces grow to something bigger or we're on a device with a slower USB connection. I can certainly add this to the commit message and/or a comment, thanks!

Patch-set: 1
Reviewer: Gerrit User 1245231 <1245231@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0