Fix loading metrics after 7c166a82
Catapult 7c166a82 changed the marker string of cache warming periods
to "telemetry.internal.warm_cache.{warm,hot}.start", but it does not
match the regexp in ChromeProcessHelper.findTelemetryInternalRanges_.
As a result, loading metrics did not ignore cache warming page loads.

To fix this, This patch relaxes the regexp to allow string with dots
between "telemetry.internal." and ".start".

Bug: chromium:770726
Change-Id: I0049b5b9c4787bfeba1928ee7c9242929682e9ca