Fix detection of GPU process in trace_event_importer
trace_event_importer looks at thread names to find the GPU process.
This is then later used as the preferred source for frame timestamp
data (although there are fallbacks to other events which can provide
less accurate versions of this data).

This patch enables telemetry to pick up the GPU process when the
DrmThread is present in the trace, but the CrGpuMain thread is not,
which then enables it to use DrmEventPageFlipComplete.

Workloads like top_25_smooth generate events in the CrGpuMain thread
so are not affected, but similar workloads (e.g. which inherit from
Smoothness as per top_25_smooth) which display a static page without
scrolling do not generate such events and therefore generate less
accurate results.

This significantly improves accuracy of percentage_smooth metrics
for this kind of workload.

Signed-off-by: Dave Rodgman <dave.rodgman@arm.com>
Bug: catapult:#3941
Change-Id: I94045fe29819920b024857a0751cd093a89f010e