Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think this is blocked on us making sure catapult/third_party/snap-it the source of truth?

pdr@ was sitting next to me yesterday when we discussed on that other change I sent. I think he's open to whatever we all feel like doing. Currently we thought making catapult canonical was a good idea.

Do we need to move it to catapult/third_party/snap-it first? I can send a change for that if so, and rework this and the other changes to move them there. But we should be able to do the code review now anyway, right?

Patch-set: 1
Reviewer: Gerrit User 1128438 <1128438@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>