Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> How come we don't get any trace events on the CrGpuMain thread if there's an animation running (the bug report mentions leaves falling)?
> 
> Just want to make sure this isn't papering over another issue.

If scrolling takes place, the only event in CrGpuMain is a regular (once-per-frame) flow event (with categories "input" and "benchmark") linked to an earlier event in CrBrowserMain. This isn't seen for non-scrolling benchmarks.

I think that this flow event is related to the scrolling input events - without scrolling, there are no input events, so no corresponding flow events.

So if there is a larger issue, it's perhaps that the mechanism by which events within the GPU process are picked up is a bit brittle. This patch does address the case for DrmEventFlipComplete, but potentially a similar issue could arise if a metric wanted to pick up an event from some other thread in the GPU process, and if neither CrGpuMain or DrmThread were present.

The only other thread which seems potentially of interest is TaskSchedulerContentChildForegroundBlockingWorker, which (I think) is involved in swapping GL buffers - perhaps this should be added to the list of thread names?

Patch-set: 3
Reviewer: Gerrit User 1239517 <1239517@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>