Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > Hey, unfortunately my plane ride last week did not allow for much time to hack on flot->d3.
> > 
> > But I'm going to try and hack on it in my spare time and ask questions along the way.
> > 
> > First question is super basic in this CL: when I run this dev_server test, I see an SVG element that is 0x0 even though I specified width 600, height 400: https://screenshot.googleplex.com/uT8sddZpkkE
> > Anyone know the right way to ensure the proper width/height? Tried various incantations of width/height CSS on the SVG element and containing divs in devtools to no avail.
> 
> Sure, feel free to ask any questions! When I started with d3 I found it super confusing at first but now feel a lot more comfortable with it. It's like the exact opposite of Polymer -- while Polymer's data binding does a lot of "spooky action at a distance", d3 is very explicit.

Thanks! Yep, that was it. The original code created the element, set properties and then added it to the DOM, but adding it to DOM and then setting properties seems to fix the sizing issues. Updated this to be a working patch, but will keep hacking locally a bit before I decide if it's worth checking in.

Patch-set: 2
Reviewer: Gerrit User 1001913 <1001913@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>