Allow results.html to use either location.search or location.hash.
I had assumed that query parameters would survive when cloudstorage
redirects to apidata.googleusercontent.com, but that does not appear
to be the case.
This CL makes results.html use location.hash if that is already set,
or location.search if location.hash is not already set.

This blocks "Set reference column to TOT in HTML Results link."
https://chromium-review.googlesource.com/c/chromium/tools/build/+/695806

http://www/~benjhayden/711062.html#s=count

https://github.com/catapult-project/catapult/issues/3909

Change-Id: I84cbeb83224acb0776e09b762fdc1a1cf030616a