Ensure snapshot-path ends with .html suffix.
Precursor to using the suffix-stripped snapshot-path as a directory
into which external images will be fetched.

We could alternately, or in addition, add an option to specify the
directory for external image storage, but it seems like the user
should always write output to a file ending with .html in any case.

Bug: chromium:762705
Change-Id: I6e24792296e4362cf6540bb3496aa71187d75e8e