Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The CL has been updated. PTAL. There seems to be no good way to issue warnings, so I'm throwing errors when unexpected seek-related events are seen.

I don't think throwing errors is a good idea because we may want to have some scenarios where seek time is not calculated but other things (like memory usage, power, time-to-play, etc.) can still be calculated. It is best if when a SWE adds a scenario like that they don't have to go edit this code. (I want to keep the barrier-to-entry low so that SWEs will use it.) Please work with tracing team to see whether or not there is a way to have warnings.

Patch-set: 2
Reviewer: Gerrit User 1149061 <1149061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>