Dashboard - Reuse commit position resolution already in pinpoint_request.
Initially pinpoint_request required you to specify git hashes, so there
was code in here to resolve to the git hash from the commit position.
We've since added support for both git_hashes or commit positions in
pinpoint_request, so we can remove the code here since it duplicates
effort.

Bug: catapult:#3959
Change-Id: Ie99514dd547fdeab3ff73d9fada3d9a1b7595dc3