Simplify histogram-span API.
Histogram-span currently supports setters for histogram and referenceHistogram
so, if a caller wants to set both, there is no way to avoid the unnecessary
updateContents_ between them. The API pattern also doesn't fit with the other
dom-modules, which all support a single build() method that can take multiple
parameters.
This CL refactors histogram-span's histogram and referenceHistogram setters to a
build() method to match the other dom-modules.

Change-Id: I12d7a3aa59358ffdd112c41636fc47d9b2a095e4