Improve seek time histogram in media metrics
The seek time histogram in previous implementation reported very
different value from the non-TBMv2 benchmark, and is not indicative of
the seek time perceived by the user. This change adds a new seek time
histogram that more closely matches the delay perceived by the user,
and renamed the previous seek time to pipeline seek time.

Bug: chromium:700160
Change-Id: I96bfadb38405417fa599e7b4170da17428b8b02e