Introduce skip files to telemetry_unittests
Currently to skip a test in telemetry_unittests one needs to either use the
@decorators method, which does not have the granularity needed for certain
configurations. Or one has to use --skip

The --skip method doesn't scale well for running one trybots as it leads to the
configuration growing.

This change introduces a filter file to define a set to skip. This is
based on the filter files already used in chromium to configure trybots.
For these filter files each line can contain an expression to define 
tests to skip. Similar to just appending a series of --skip to the 
command line.

TEST: ran telemetry_unittests locally with a skip file
Bug: chromium:758065
Change-Id: I0fa4634c20860c18158e55df64496d59204d8168