Update patch set 5
Patch Set 5:

It looks like this CL adds:

1) Scripts to vulcanize histograms_viewer.html
2) A checked-in version of histograms_viewer.html

I don't see any guarantee that the checked-in version of histograms_viewer.html will stay up to date. I'm sure we could add a presubmit, but instead why not generate this as part of the process to build/deploy pinpoint? Long-term, we'd like to do this for perf dashboard components too, but I think that's also blocked on the vulcanizer supporting polymer elements?

Patch-set: 5
CC: Gerrit User 1001913 <1001913@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>