Save sample values alongside sample diagnostics.
Currently, Histogram stores all sample values in a single array to enable MWU
hypothesis test. It stores sample diagnostics in arrays in HistogramBins. There
is no way to recover the exact sample values that originally accompanied sample
diagnostics.
Users of results.html need to be able to see the sample values that went with
the sample diagnostics.
This CL formalizes Sample = value + diagnostics.

https://github.com/catapult-project/catapult/issues/2848
https://github.com/catapult-project/catapult/issues/3972

Change-Id: I0614d6c0bcbc06b067cf3ac2a99f1b7c30bbf25a