Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > I think this change should be made to typ (https://github.com/dpranke/typ/). We are in the middle of designing a unified test disabling file format, so please stay tune.
> > 
> > Do you know when you plan to switch over to this planned format?
> 
> +Charlie & Randy will be driving this effort. I would get in touch with them.
> 
> Mean while the test configuration skip work around still work for you, right? If that list of disabled tests keep growing, I think someone from ChromeOS team should look into fixing the tests

For telemetry benchmarks, we plan on refactoring the recently refactored test disabling in telemetry (go/storyexpectationsDD) to use the new format. This is tangential to this work though, since this appears to be for unit tests and not benchmarks. Telemetry benchmarks were the only thing moved to StoryExpectations. The telemetry unit tests still use the legacy way to disable (mostly using decorators).

Patch-set: 4