Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > PTAL
> 
> Sorry, what's the delta from the previous patch?

So my feedback yesterday during our 1:1 was to largely in line with what you suggested (making this a deploy step for Pinpoint instead of checking it in). Suggestion was to strip this down to just what Pinpoint needed without the changes to how Telemetry generates it's html results file, and we can integrate it into Pinpoint's deployment. This looks like it does that, but Ben to confirm.

+1 to better description

Patch-set: 7
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>