NOT FOR REVIEW - Testing addSeriesGroup removal.
Seem to be 2 main callers of addSeriesGroup left:

chart-container:onTitleClicked: Which modifies the path (depending on where you clicked) of the
chart. All the new test paths added from that are unselected.

group-report-page:addGraph: Which take the path from an alert (and the ref is possible), and adds
a chart with those 2 selected and the rest unselected.

I'm not sure it's possible for either of these to know ahead of time all the test paths, so they
need to query into /list_tests.

Demo:

https://dev-simonhatch-1ddd6778-dot-chromeperf.appspot.com/report?sid=5628cd6205c3822e07dc543248dbf926ce2250bf75e4345d0e63157e87bfe95b
https://dev-simonhatch-1ddd6778-dot-chromeperf.appspot.com/group_report?keys=agxzfmNocm9tZXBlcmZyFAsSB0Fub21hbHkYgIDg66_ItwkM

Bug: catapult:#3272
Change-Id: I2402d97a8d5de0bd1bc8f798f4fc66dc6063c4c7