Update patch set 5
Patch Set 5:

Hey, had some time so took a quick stab at this. It's not really ready for a real review, wanted to check if it was functionally correct and start some discussion on how to finish it.

Patch-set: 5
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false