Serialize selected related timeseries tab in short uri.
Chart state is currently serialized as an array of series groups.
This array does not permit storing other kinds of state for each chart.
This CL changes chart state to a dictionary containing the seriesGroups array
plus optionally a selectedRelatedTab string.
We can add other fields to this dictionary in the future if necessary.

https://dev-benjhayden-16849b74-dot-chromeperf.appspot.com/report?3405=y&sid=98d70d00c007c46f091c75795164ce1490c1efd78b8ef124da43585752a0a41f

https://github.com/catapult-project/catapult/issues/3852
Bug: catapult:#3852

Change-Id: I9903d7a14fa361afe70102163e41435b90cead41