Update 'Date' as current date in response header in WPR-go server
In WPR-go server, the 'Date' in response header is the date and time the
response was archived in the wprgo file, rather than the date it was
generated and sent from the WPR-go server. This would make some resources
to be requested again, instead of loading from cache directly in warm
cache case, which may impact user experience and browser performance.

This CL follows the HTTP spec
(https://tools.ietf.org/html/rfc7231#section-7) and updates the 'Date'
attribute as 'current time' in WPR-go server.

Bug: chromium:#760422
Change-Id: I479beff89d67e98c16189847aa9855856130b23c