Dashboard - Send chart_name and statistic separately in request to Pinpoint.
This is required in order for Pinpoint to use histogram output. In
chartjson, a metric like 'cpu_time_avg' is directly addressable in
the chartjson results, whereas in histogramset would store statistics
like this in the RunningStatistics member of the histogram. So you'd
have to look up the histogram without the 'avg', and then grab the
mean or whichever statistics property you want to bisect on.


https://github.com/catapult-project/catapult/issues/3986
Bug: catapult:#3986
Change-Id: I6b07883565990ceef1fc2ed3b9c1144961fc7c50