Make python Histograms and Diagnostics recognize longs as numbers.
The python implementation of the Breakdown diagnostic only recognizes ints and
floats (including NaN and +/- Infinity) as numbers, and throws an AssertionError
for any other value type.

This AssertionError was raised on a windows bot for the multitab:misc:typical24
story:
https://chromium-swarm.appspot.com/task?id=395074104c2bbf10&refresh=10&show_raw=1

The AssertionError doesn't say what the invalid value was, but it must have been
something that passed the corresponding type check in the javascript
implementation of Breakdown (typeof value !== 'number'), and was successfully
serialized as JSON in vinn and deserialized in Python. The only thing I can
think of that fits these criteria is a python long. That suggests that the
breakdown entry value was greater than 4GB, which a 32-bit build of python would
deserialize as a long. I can easily believe that opening 24 tabs would consume
more than 4GB of memory.

This CL adds long to the list of acceptable Breakdown values in the python
implementation and wraps the type check in a function IsNumberType. There are a
couple other similar type checks, so they are also updated to use IsNumberType().

A test is added to ensure that Breakdown can handle longs, which reproduces the
error seen on the bot without this fix.

I added error messages to the other assertions in this file so we don't need to
guess at how they failed.

Bug: chromium:776608
Change-Id: I8b399dd277e89f623437bef757b0e19611b6f27b