Pinpoint - Add histogram output support.
This is the next step in supporting perf try jobs, we need histogram
output support since we'll need to plumb that through to the
vulcanized results2 eventually.

This adds a read_value.ReadHistogramsJsonValue quest, similar to
the ReadChartJsonValue and ReadGraphJsonValue quests that already exist.
As described in the bug, there's an issue with getting rid of the
existing ReadChartJsonValue class right away, as older jobs will
fail during unpickling and we'll need to figure out some migration
steps there.

https://github.com/catapult-project/catapult/issues/3986
Bug: catapult:#3986
Change-Id: I9fceb385dc801ea29be5f56c3fc29280bb004b9d