Pinpoint - Results2 endpoint for perf try jobs.
This adds a working results2 endpoint for perf try jobs. Takes as a
url parameter the job_id, pulls the data from the job, parses out the
isolates, fetches the histogram data, and renders the vulcanized
viewer + the histogram data.

As part of the deployment, the dashboard will run the vulcanize and
upload the viewer so that Pinpoint can read it.

Bug: catapult:#3986
Change-Id: Ic2633fb8f534b737237b4cdc9b17eff5dab7ef03