Pinpoint - Results2 endpoint for perf try jobs.
This adds a working results2 endpoint for perf try jobs. Takes as a
url parameter the job_id, pulls the data from the job, parses out the
isolates, fetches the histogram data, and renders the vulcanized
viewer + the histogram data.

As part of the deployment, the dashboard will run the vulcanize and
upload the viewer so that Pinpoint can read it.

After this, we need get a UI up to kick off perf try jobs and surface
the results2 links on the job page once it's complete.

https://github.com/catapult-project/catapult/issues/3985
Bug: catapult:#3985
Change-Id: Ic2633fb8f534b737237b4cdc9b17eff5dab7ef03