DiscoverModule: just log ERROR for unable-to-import modules.
Sometimes Telemetry fails to run because DiscoverClasses raises an
uncaught ImportError when it scans a ill-behaved module. Reasons
of an ill-behaved a module could be either syntax error, importing
modules not in the platform environment, etc.

It is troublesome that a Telemetry task (e.g. benchmark_runner) is
affected by unrelated module(s) (say another benchmark totally
irrelevant to the target benchmark).
See https://crbug.com/756053 for example.

BUG=777759

Change-Id: I99289b64efc3deb30d7b219dff6b9118e8ade16b