Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why do you need this?

In this particular case, to run the smoothness gpu rasterization benchmarks on ChromeOS devices. 

I believe the alternative would be to add 'story_module.expectations.ALL_CHROMEOS' to the 'SUPPORTED_PLATFORMS' for 'SmoothnessGpuRasterizationTop25'? However, I imagined that this would be less acceptable to submit, so I thought adding an extra flag that allows us to enable the use of the benchmark for our particular use case would be more agreeable. Why not allow the user to also override the platforms supported declaration, if you already allow them to override the @Disabled property?

In either case, it's confusing if it currently fails because the platform isn't in SUPPORTED_PLATFORMS, but the error message invites you to use --also-run-disabled-tests which doesn't actually help.

Patch-set: 1
Reviewer: Gerrit User 1190714 <1190714@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0