Update patch set 1
Patch Set 1:

A note of verbage:

what about calling them "playbacks". So then for this cl description we can say something like

"
Update media metric to support multiple *playbacks* in a trace.

Make use of the *playback* id that has been added to the media trace events to support multiple meida playbacks in a single trace.
"

The reason this is important is the the id relates to the playback, not to the media file itself. So I can play BigBuckBunny.mp4 multiple times and each time I get a new playback_id. 

Not sure whether or not this needs to be updated in the other CL as well.

Patch-set: 1
Reviewer: Gerrit User 1149061 <1149061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>