viz: Fix collecting rendering stats with viz.
With viz, the display compositor lives in the gpu process. So it
is possible for the gpu process to have rendering stats. So update
the code to handle that case.

Also, with the mojo viz service, the thread is called 'VizMain'.
So check for a thread with that name to determine if it's the gpu
process.

BUG=chromium:779316

Change-Id: Id95368745942978b4e5233d264af727c6b1c6c90