Syntax updates for Polymer 2.0.
This CL is moved from a pull request:
https://github.com/catapult-project/catapult/pull/4016

Polymer 2.0 is production ready, and we advise all projects to begin migration.
To aid in that migration, we're writing and applying some automated fixes for
the mechanical parts of the upgrade.

The <content> element is deprecated, part of the Shadow DOM v0 API which has
been supplanted by Shadow DOM v1's <slot> element. For very simple uses – like
those in this PR – this change should be safe and will not require changes to
the usage sites of elements. Therefore, these changes are backwards-compatible
with Polymer 1.0 applications.

As <content> is the more powerful API, Polymer 1.0 supports both <content> and
<slot> by converting <slot> back into <content> at runtime. Polymer 2.0 only
supports <slot>. There will likely never be cross-browser native support for
<content>, while Safari and Chrome already support <slot> natively.

This change also updates usage of `@apply` to remove the parentheses. Like
the other changes in this CL, these are backwards compatible with Polymer 1.0,
and necessary for Polymer 2.0.

Screenshot of the trace viewer with some events selected, the metrics side
panel open, and its Export dropdown open:
https://imgur.com/arvWmUx

Change-Id: Iba230c9af872d02eaabb58279da06492e8d58794