Fix typo in histogram-set-controls.
The array of possible value for the statistical significance threshold `alpha`
is computed to be [0.001, ..., 0.01, ..., 0.09, 0.01].
The last entry is the problem.
When the user tries to set alpha to the maximum possible value, the slider
unexpectedly jumps from the top back to the middle because the value 0.01 is in
the array twice.

This CL fixes the last entry to be 0.1.

Change-Id: I9288569583a0cd0247ab5337e8ae5e0b39869bca