Allow users to control which statistics are displayed in results.html.
Results.html allows users to select a single statistic to display in
a scalar-span in the histogram-set-table-cells when they are closed.
The name of the selected statistic is communicated across results.html
as HistogramSetViewState.displayStatisticName.

When a cell is open, that scalar-span is hidden and the histogram-span
is shown. The statistics displayed in histogram-span are controlled by
the metric using Histogram.customizeSummaryOptions(). These summary
options also control which statistics are uploaded to the chromeperf
dashboard.

However, sometimes users want to see multiple statistics when cells
are closed. Also, users sometimes want to see different statistics
in open cells than what the metric wants to export to the chromeperf
dashboard.

This CL replaces the statistic selector in histogram-set-controls
with a dropdown containing 2 grids of toggle buttons to allow users
to select multiple statistics to display when cells are closed and open.

The string HistogramSetViewState.displayStatisticName is replaced
with 2 arrays of strings: openCellStatisticNames and
closedCellStatisticNames.

When there are multiple closedCellStatisticNames,
histogram-set-table-cell displays a scalar-map-table instead of the
scalar-span.

histogram-span displays openCellStatisticNames in its scalar-map-table
instead of the summary options from the metric.

Demo: http://www/~benjhayden/731652.html

https://github.com/catapult-project/catapult/issues/3380
Bug: catapult:#3380

Change-Id: I47e9bf8957e5df4a2481d416806aa1c290440230