Update patch set 4
Patch Set 4:

This is really cool. I'm wondering if we should use metaprogramming to make sure that _all_ instances of Diagnostic use __slots__ in the right way. The argument in favor would be that it would eliminate a class of mistakes (people not using __slots__) but maybe it's a premature abstraction. Thoughts?

Patch-set: 4
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>