Fix a bug with results.html deep-linking sortDescending.
One of the pieces of ViewState that is serialized in results.html URLs is a flag
whether the sorting column is ascending or descending. The query parameter for
this is 'd'. The ViewState property is 'sortDescending'. When '?d=' is in a
results.html URL, the sorting column should be descending, but this is currently
broken. HistogramSetLocation is doing its job, the bug is in histogram-set-table.

h-s-t.updateContents_() always sets sortColumnIndex and sortDescending on its
tr-ui-b-table. When HistogramSetLocation restores them from the URL, h-s-t first
sets sortColumnIndex, which causes tr-ui-b-table to dispatch a
sort-column-changed event, which causes h-s-t to update both sortColumnIndex and
sortDescending in its ViewState, which causes the ViewState to call back into
h-s-t.updateContents_(). The sortDescending setting is lost in the reentrant
cross-talk.

This CL makes histogram-set-table.onSortColumnChanged_ ignore the
sort-column-changed event when it was fired during h-s-t.updateContents_().
onSortColumnChanged_ can detect this situation without a dedicated flag because
it knows that h-s-t.updateContents_() always sets this.tableRows_, then
sortColumnIndex and sortDescending, then this.$.table.tableRows, so if
table.tableRows !== this.tableRows_, then it is being called during
updateContents_(), not in response to a user gesture, so it can ignore the event.
This is definitely weird, and it could probably be avoided with Polymer-Redux...

Change-Id: I4ac3d7fc6bcc5c9bbfc423c256fb9dfff3de8be4