Fix a bug with results.html deep-linking sortDescending.
One of the pieces of ViewState that is serialized in results.html URLs is a flag
whether the sorting column is ascending or descending. When '?d=' is in a
results.html URL, the sorting column should be descending, but this is currently
broken. HistogramSetLocation is doing its job, the bug is in histogram-set-table.

h-s-t.updateContents_() always sets sortColumnIndex and sortDescending on its
tr-ui-b-table. When HistogramSetLocation restores them from the URL:
- h-s-t sets tr-ui-b-table.sortColumnIndex,
- which causes tr-ui-b-table to dispatch a sort-column-changed event,
- which causes h-s-t.onSortColumnChanged_ to update both sortColumnIndex and
sortDescending in its ViewState,
- which causes the ViewState to reenter h-s-t.updateContents_().
The sortDescending setting is lost in the reentrant cross-talk.

This CL makes updateContents_() temporarily stop listening for the
sort-column-changed event while it updates sortColumnIndex and sortDescending
in order to prevent the listener from unnecessarily updating viewState.

Change-Id: I4ac3d7fc6bcc5c9bbfc423c256fb9dfff3de8be4