[pinpoint] Separate exception comparisons for each Quest.
Previously, the bisect comparison just compared the Attempt statuses. But that
hides knowledge of which Execution failed. If different Executions fail, they
should be treated as different failures for bisection.

Change the properties exposed by Attempt to provide the ones we need to dig into
the Execution data. The new properties better align with the Attempt's internal
structure anyway.

Bug: catapult:#4025
Change-Id: I7905e3384e8ddaa5a14bbe25963f26349fee1caf