Factor EmptyHistogramBin out of HistogramBin to save memory.
Now that HistogramBins are copy-on-write, it is guaranteed that the
HistogramBins created by HistogramBinBoundaries will always be empty.
Their count and diagnostic_maps properties will always waste space.

This CL reclaims that space by factoring EmptyHistogramBin out of
HistogramBin. EmptyHistogramBin contains the _range property, but uses
static getters for count and diagnostic_maps. It supports AsDict()
but not FromDict().

Change-Id: I98156ef6c03ad39ff3b6ae2269370d10f43bec07