Update patch set 33
Patch Set 33:

> Patch Set 33:
> 
> (2 comments)
> 
> I also think heap_profiler should not be part of the tracing/value/ module.
> 
> Maybe tracing/utils/ ?

I don't want to create tracing/utils/ just for the purpose of storing this file because we already know that we want to move tracing/value/ out of tracing and reorganize it, and tracing/utils/ would be more mess to clean up. Unfortunately, this is a difficult problem because it also depends on tracing/base/ and that prompts broader questions about the state of Catapult's directory structure.

The reason tracing/value/ is the best possible place for this file right now is that it's the only sensible place in tracing/ that is already a Python module. I agree that it's not perfect, but I would like to compromise with an eye toward working out the directory structuring issues.

Ned, I sat down with Ben to look at some of the issues we might have to solve in order to get tracing/base/ out of tracing and we came up with this doc: https://docs.google.com/document/d/1XCYoDqMAI6UUX8ZixtQtLwSaQjxuNcN64QiJ-tTjB58/edit. There are a bunch of questions and problems to address in it, and I was thinking maybe you, me, and a third person, maybe Annie, should sit down and talk about those when I'm in Michigan next week. WDYT?

Patch-set: 33
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>